Fixing some typos

+++ test/misc/addresssettings/addresssettings_redeliver_test.go
@@ -53,7 +53,7 @@ var _ = ginkgo.Describe("AddressSettingsRedeliveryTest", func() {
 		urls, err := brokerDeployer.GetExternalUrls(ExpectedURL, 0)
 		address := urls[0]
 		value := retrieveAddressSettings(address, AddressBit, hw)
-		gomega.Expect(value.RedeliveryMultiplier).To(gomega.Equal(1.0))
+		gomega.Expect(value.RedeliveryMultiplier).To(gomega.Equal(float32(1.0)))
 	})
 
 	ginkgo.It("RedeliveryDelay check", func() {
